StoryRag/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pypdf>=5.0.0
# Use <=1.3.2 for broad Windows compatibility
scikit-learn>=1.3.0,<=1.3.2
numpy>=1.24.0
fastapi>=0.115.0
uvicorn>=0.30.0
langdetect>=1.0.9

